Application with errors
I am going to write the same things on Slack tomorrow morning.
Throughout the day I try to figure it out how to have another webpage, it is not very complicated but I am getting some troubles now.
The best option was to make a new project using the Vue UI to include the vue router, because we don't have one. I tried to install that using the webstorm terminal but when I try to use that our template is not working anymore. 
In this version I manually added the router but I am missing something somewhere